Refactor ContestProblems into it's own component
