+++ packages/app/typings/global.d.ts
@@ -1,3 +1,5 @@
+declare type ProblemStatus = "not submitted" | "accepted" | "rejected";
+
 declare interface ProblemOverview {
   id: string;
   origin: string;
